refactor(routes): split task-workflow routes out of tasks router (SHR-177)
Move board/workflow/metadata sub-routes (move, summary, note, refs, updates, hooks)
from tasks.ts into new task-workflow.ts. Extract validateCreateTaskBody,
applyDraftUpdates, and resolveTaskTitleAndDescription helpers into _shared.ts.
tasks.ts is now 384 lines (down from 774); task-workflow.ts is 248; all files under ~400.
